Refactor: move testing inventory command to dedicated package
